Python updates: - Added an input_source enum that represents using a local FASTA file as input. - Structured the code for use as a module.
