--- scripts/run_treesort.py
@@ -1,236 +1,0 @@
-
-import argparse
-from common import DEFAULT_REF_SEGMENT, InferenceType, InputSource, InputParameter, INVALID_FASTA_CHARS, Method, safeTrim, VALID_SEGMENTS
-from enum import Enum
-import json
-import re
-import os
-import shutil
-import subprocess
-import sys
-
-
-# TODO: use https://github.com/BV-BRC/bvbrc_subspecies_classification/blob/master/scripts/run_subspecies_classification.py as an example.
-
-
-class TreeSortRunner:
-
-   # The base URL
-   base_url: str = None
-
-   # The input filename with full path.
-   input_filename: str = "input.fasta"
-
-   # The JSON data for the job.
-   job_data: dict = None
-
-   # The output filename with full path.
-   output_filename: str = None
-   
-
-   # C-tor
-   def __init__(self, job_data, output_filename):
-         
-      # Set and validate the job data.
-      self.job_data = job_data
-      if not self.job_data:
-         raise ValueError("Invalid job data")
-      
-      # Set and validate the output filename.
-      self.output_filename = safeTrim(output_filename)
-      if self.output_filename == "":
-         raise ValueError("Invalid output filename")
-
-      # Determine the base URL.
-      if "P3_BASE_URL" in os.environ:
-         self.base_url = os.environ["P3_BASE_URL"]
-      else:
-         self.base_url = "https://www.bv-brc.org"
-         
-      
-
-   @staticmethod
-   def is_job_data_valid(job_data: dict) -> bool:
-
-      try:
-         if not job_data or not isinstance(job_data, dict) or len(job_data) == 0:
-            raise Exception("job_data is empty")
-         
-         # Validate the reference segment and provide a default if not provided.
-         refSegment = safeTrim(job_data[InputParameter.RefSegment])
-         if not refSegment:
-            refSegment = DEFAULT_REF_SEGMENT
-
-         if not refSegment in (segment.value for segment in VALID_SEGMENTS):
-            raise ValueError(f"Invalid reference segment: {refSegment}")
-         
-         # Validate the segments
-         segments = safeTrim(job_data[InputParameter.Segments.value])
-         for segment in segments.split(","):
-            if not segment in (segment.value for segment in VALID_SEGMENTS):
-               raise ValueError(f"Invalid segment: {segment}")
-            
-      except Exception as e:
-         sys.stderr.write("Job data is invalid:\n %s" %(e))
-         return False
-
-      return True
-   
-
-   # Build alignments and trees and compile a descriptor file.
-   def prepare_dataset(self) -> bool:
-
-      try:
-         # Example usage: ./prepare_dataset.sh --segments "HA,NA" segments.fasta HA myoutdir
-         cmd = ["prepare_dataset.sh"]
-
-         inference_type = job_data[InputParameter.InferenceType.value]
-         if (inference_type and inference_type == InferenceType.FastTree.value):
-            cmd.append("--fast")
-
-         segments = job_data[InputParameter.Segments.value]
-         if segments:
-            # TODO: validate segments
-            cmd.append("--segments")
-            cmd.append(segments)
-
-         cmd.append(self.input_filename)
-
-         refSegment = job_data[InputParameter.RefSegment.value]
-         if refSegment:
-            # TODO: validate refSegment
-            cmd.append(refSegment)
-
-         cmd.append(self.output_filename)
-
-         # TODO: uncomment
-         # result = subprocess.check_call(cmd, shell=False)
-         print(cmd)
-
-      except ValueError as e:
-         sys.stderr.write("Error preparing dataset:\n %s" %(e))
-         return False
-         
-      return True
-     
-
-   # Determine the input source and process the input file accordingly.
-   def process_input_file(self) -> bool:
-
-      try:
-         # The input source determines how the input file is provided.
-         input_source = job_data[InputParameter.InputSource.value]
-
-         if input_source == InputSource.FastaFileID.value:
-
-            # Fetch the input file from the workspace.
-            try:
-               fetch_fasta_cmd = ["p3-cp", "ws:%s" %(job_data[InputParameter.InputFastaFileID.value]), self.input_filename]
-               subprocess.check_call(fetch_fasta_cmd, shell=False)
-
-            except Exception as e:
-               raise RuntimeError("Error copying FASTA file from workspace:\n %s" %(e))
-
-         elif input_source == InputSource.FastaData.value:
-         
-            # Copy user data to the input file.
-            try:
-               with open(self.input_filename, "w+") as input_file:
-                  input_file.write(job_data[InputParameter.InputFastaData.value])
-
-            except Exception as e:
-               raise IOError("Error copying FASTA data to input file:\n %s" %(e))
-
-         else:
-            raise ValueError(f"Invalid input source: {input_source}")
-
-         # Validate the input file.
-         if not os.path.exists(self.input_filename) or os.path.getsize(self.input_filename) == 0:
-            raise IOError("Input FASTA file is invalid or empty")
-
-         # Remove invalid characters from FASTA headers.
-         with open(self.input_filename, 'r+') as f:
-            data = ""
-            for line in f:
-               if line.startswith(">"):
-                  line = line.strip()
-                  line = re.sub(INVALID_FASTA_CHARS, "", line)
-                  line = re.sub(" ", "_", line)
-                  line += "\n"
-
-               data += line
-
-            f.seek(0)
-            f.write(data)
-            f.truncate()
-
-      except Exception as e:
-         sys.stderr.write("Error processing input file:\n %s" %(e))
-         return False
-      
-      return True
-
-
-if __name__ == "__main__" :
-  
-   parser = argparse.ArgumentParser(description="A script to run TreeSort")
-   parser.add_argument("-j", "--jobFilename", dest="job_filename", help="A JSON file for the job", required=True)
-   parser.add_argument("-o", "--outputDirName", dest="output_dir_name", help="The output directory name (defaults to current directory)", required=False, default=".")
-
-   args = parser.parse_args()
-
-   # Validate the job filename
-   job_filename = safeTrim(args.job_filename)
-   if len(job_filename) == 0:
-      sys.stderr.write("Invalid job filename parameter")
-      sys.exit(-1)
-
-   # Load job data
-   job_data = None
-   try:
-      with open(job_filename, "r") as job_file:
-         job_data = json.load(job_file)
-
-   except Exception as e:
-      sys.stderr.write("Invalid job file:\n %s" %(e))
-      sys.exit(-1)
-
-   # Validate the job data
-   if not TreeSortRunner.is_job_data_valid(job_data):
-      sys.exit(-1)
-
-   # For debugging   
-   #print(job_data)
-
-
-   # Validate the output directory name
-   output_dir_name = safeTrim(args.output_dir_name)
-   if len(output_dir_name) == 0:
-      sys.stderr.write("Invalid output directory name parameter")
-      sys.exit(-1)
-
-   # Create the output directory if it doesn't exist.
-   output_dir_name = os.path.abspath(output_dir_name)
-   if not os.path.exists(output_dir_name):
-      os.mkdir(output_dir_name)
-
-   # Go to the output directory.
-   os.chdir(output_dir_name)
-
-   # Create a TreeSortRunner instance
-   runner = TreeSortRunner(job_data, output_dir_name)
-
-   # Process the input file
-   runner.process_input_file()
-
-   # Prepare the dataset
-   runner.prepare_dataset()
-
-   # TODO: Does this need to end in .txt?
-   #output_filename = os.path.join(output_dir_name, job_data["output_file"] + ".txt")
-
-   # Define the input file
-   #input_filename = os.path.join(output_dir_name, "input.fasta")
-
-
-
