+++ scripts/treesortrunner/__init__.py
@@ -1,0 +1,1 @@
+from .run_treesort import main
