+++ test/test_job.json
@@ -1,0 +1,21 @@
+{
+   "clades_path": "",
+   "descriptor_path": "",
+   "deviation": 2.0,
+   "equal_rates": true,
+   "inference_type": "FastTree",
+   "input_fasta_data": null,
+   "input_fasta_file": "test/swine_H1_HANA.fasta",
+   "input_fasta_file_id": "TODO",
+   "input_source": "fasta_file",
+   "is_time_scaled": false,
+   "match_on_epi": false,
+   "match_on_regex": null,
+   "match_on_strain": false,
+   "method": "local",
+   "no_collapse": true,
+   "output_path": "",
+   "p_value": 0.001,
+   "ref_segment": "HA",
+   "segments": null
+}
